requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==21.2.0